chore: renames components
BREAKING CHANGE: FLIPMove has been renamed to Move. Target has been renamed to FocalTarget.
