+++ packages/yubaba/src/Baba/test.tsx
@@ -2,7 +2,7 @@ import * as React from 'react';
 import { mount } from 'enzyme';
 import 'jest-enzyme';
 import { Baba } from '../Baba';
-import Target from '../Target';
+import Target from '../FocalTarget';
 import { getElementSizeLocation } from '../lib/dom';
 import defer from '../lib/defer';
 import * as utils from '../__tests__/utils';
