+++ packages/yubaba/src/index.tsx
@@ -7,8 +7,7 @@ export { default as Wait } from './Wait';
 export { default as Noop } from './animations/Noop';
 export { default as CrossFadeMove } from './animations/CrossFadeMove';
 export { default as FadeMove } from './animations/FadeMove';
-export { default as Move } from './animations/CrossFadeMove'; // To be replaced with animations/Move
-export { default as FLIPMove } from './animations/Move'; // To be renamed to Move
+export { default as Move } from './animations/Move';
 export { default as Swipe } from './animations/Swipe';
 export { default as CircleExpand } from './animations/CircleExpand';
 export { default as CircleShrink } from './animations/CircleShrink';
@@ -19,7 +18,7 @@ export { default as ConcealMove } from './animations/ConcealMove';
 // Utility stuff
 export * from './Collector';
 export { default as Collector } from './Collector';
-export { default as Target } from './Target';
+export { default as FocalTarget } from './FocalTarget';
 export * from './lib/curves';
 export * from './lib/dom';
 export * from './lib/math';
